[Feat] Added dynamic fetching of terms of use page depending on different builds from env file
