+++ src/routes/_TermsOfUse.scss
@@ -26,7 +26,7 @@ main {
       font-family: "Montserrat", sans-serif;
       font-size: 16px;
       font-weight: 400;
-      color: #a0a0ac;
+      color: #c8c8c8;
       margin-bottom: 1em;
     }
   }
