+++ src/routes/terms-of-use.jsx
@@ -1,339 +1,26 @@
-import React from "react";
-import "./_TermsOfUse.scss";
+import React, { useState, useEffect } from 'react';
 
-export default function TermsOfService() {
-  return (
-    <main>
-      <div className="TermsSection">
-        <h1>Terms of Use</h1>
-        <div className="Content">
-          <div className="DescriptionContainer">
-            <h3>1. Agreement to Terms</h3>
-            <p>
-              These Terms of Use constitute a legally binding agreement made between you,
-              whether personally or on behalf of an entity (“you”) and all the
-              contributors of the Djed Alliance (“we," “us," or “our”), concerning your
-              access to and use of the Djed Stablecoin protocol, including its smart
-              contracts (the "Smart Contracts") and the djed.one web interface (the
-              “Interface” or "Site") (collectively, "Djed").
-            </p>
-            <p>
-              You agree that by using Djed, you have read, understood, and agreed to be
-              bound by all of these Terms of Use. IF YOU DO NOT AGREE WITH ALL OF THESE
-              TERMS OF USE, THEN YOU ARE EXPRESSLY PROHIBITED FROM USING DJED AND YOU MUST
-              DISCONTINUE USE IMMEDIATELY.
-            </p>
-            <p>
-              We reserve the right, in our sole discretion, to make changes or
-              modifications to these Terms of Use from time to time. We will alert you
-              about any changes by updating the “Last updated” date of these Terms of Use,
-              and you waive any right to receive specific notice of each such change.
-              Please ensure that you check the applicable Terms every time you use Djed so
-              that you understand which Terms apply. You will be subject to, and will be
-              deemed to have been made aware of and to have accepted, the changes in any
-              revised Terms of Use by your continued use of Djed after the date such
-              revised Terms of Use are posted.
-            </p>
-            <p>
-              Djed is not intended for use by any person or entity in any jurisdiction or
-              country where such use would be contrary to law or regulation or which would
-              subject us to any registration requirement within such jurisdiction or
-              country. Accordingly, those persons who choose to use the Protocol or the
-              Interface from such locations do so on their own initiative and are solely
-              responsible for compliance with local laws, if and to the extent local laws
-              are applicable.
-            </p>
-            <p>
-              All users who are minors in the jurisdiction in which they reside must have
-              the permission of, and be directly supervised by, their parent or guardian
-              to use Djed. If you are a minor, you must have your parent or guardian read
-              and agree to these Terms of Use prior to you using Djed.
-            </p>
-            <h3>2. User Representations</h3>
-            <p>
-              By using Djed, you represent and warrant that: (1) you have the legal
-              capacity and you agree to comply with these Terms of Use; (2) you are not a
-              minor in the jurisdiction in which you reside, or if a minor, you have
-              received parental permission to use Djed; (3) you will not use Djed for any
-              illegal or unauthorized purpose; and (4) your use of Djed will not violate
-              any applicable law or regulation.
-            </p>
-            <h3>3. Prohibited Activities</h3>
-            <p>
-              You may not access or use Djed for any purpose other than that for which we
-              make Djed available. Djed may not be used in connection with any commercial
-              endeavors except those that are specifically endorsed or approved by us.
-            </p>
-            <p>As a user of Djed, you agree not to:</p>
-            <ul>
-              <li>
-                Trick, defraud, or mislead us and other users, especially in any attempt
-                to learn sensitive account information such as user passwords or private
-                keys.
-              </li>
-              <li>
-                Circumvent, disable, or otherwise interfere with security-related features
-                of Djed, including features that prevent or restrict the use or copying of
-                any Content or enforce limitations on the use of Djed and/or the Content
-                contained therein.
-              </li>
-              <li>
-                Disparage, tarnish, or otherwise harm, in our opinion, us and/or Djed.
-              </li>
-              <li>
-                Use any information obtained from Djed in order to harass, abuse, or harm
-                another person.
-              </li>
-              <li>
-                Use Djed in a manner inconsistent with any applicable laws or regulations.
-              </li>
-              <li>
-                Upload or transmit (or attempt to upload or to transmit) viruses, Trojan
-                horses, or other material, that interferes with any party’s uninterrupted
-                use and enjoyment of Djed or modifies, impairs, disrupts, alters, or
-                interferes with the use, features, functions, operation, or maintenance of
-                Djed.
-              </li>
-              <li>
-                Upload or transmit (or attempt to upload or to transmit) any material that
-                acts as a passive or active information collection or transmission
-                mechanism, including without limitation, clear graphics interchange
-                formats (“gifs”), 1×1 pixels, web bugs, cookies, or other similar devices
-                (sometimes referred to as “spyware” or “passive collection mechanisms” or
-                “pcms”).
-              </li>
-              <li>
-                Interfere with, disrupt, or create an undue burden on Djed or the networks
-                or services connected to Djed.
-              </li>
-              <li>
-                Harass, annoy, intimidate, or threaten any of our contributors engaged in
-                providing any portion of Djed to you.
-              </li>
-              <li>
-                Attempt to bypass any measures of Djed designed to prevent or restrict
-                access to Djed, or any portion of Djed.
-              </li>
-              <li>
-                Copy or adapt Djed’s software in ways that violate that software's
-                license.
-              </li>
-              <li>
-                Except as permitted by applicable law, decipher, decompile, disassemble,
-                or reverse engineer any of the software comprising or in any way making up
-                a part of Djed.
-              </li>
-              <li>
-                Except as may be the result of standard search engine or Internet browser
-                usage, use, launch, develop, or distribute any automated system, including
-                without limitation, any spider, robot, cheat utility, scraper, or offline
-                reader that accesses Djed, or using or launching any unauthorized script
-                or other software.
-              </li>
-            </ul>
-            <h3>4. Submissions</h3>
-            <p>
-              You acknowledge and agree that any questions, comments, suggestions, ideas,
-              feedback, or other information regarding Djed ("Submissions") provided by
-              you to us are non-confidential and shall become our sole property. We shall
-              own exclusive rights, including all intellectual property rights, and shall
-              be entitled to the unrestricted use and dissemination of these Submissions
-              for any lawful purpose, commercial or otherwise, without acknowledgment or
-              compensation to you. You hereby waive all moral rights to any such
-              Submissions, and you hereby warrant that any such Submissions are original
-              with you or that you have the right to submit such Submissions. You agree
-              there shall be no recourse against us for any alleged or actual infringement
-              or misappropriation of any proprietary right in your Submissions.
-            </p>
-            <h3>5. Site Management</h3>
-            <p>
-              We reserve the right, but not the obligation, to: (1) monitor Djed for
-              violations of these Terms of Use; (2) take appropriate legal action against
-              anyone who, in our sole discretion, violates the law or these Terms of Use,
-              including without limitation, reporting such user to law enforcement
-              authorities; and (3) otherwise manage Djed in a manner designed to protect
-              our rights and property and to facilitate the proper functioning of Djed.
-            </p>
-            <h3>6. Privacy Policy</h3>
-            <p>
-              We care about data privacy and security. Djed does not collect any personal
-              data. Transactions interacting with the Smart Contracts are publicly
-              registered on the blockchain and the wallet addresses and the balances on
-              these addresses are visible. By using Djed, you agree that you understand
-              how such data is stored on blockchains and you accept the storage of such
-              data on the blockchain.
-            </p>
-            <h3>7. Term and Termination</h3>
-            <p>
-              These Terms of Use shall remain in full force and effect while you use Djed.
-              WITHOUT LIMITING ANY OTHER PROVISION OF THESE TERMS OF USE, WE RESERVE THE
-              RIGHT TO, IN OUR SOLE DISCRETION AND WITHOUT NOTICE OR LIABILITY, AND TO THE
-              EXTENT THAT IT IS TECHNOLOGICALLY FEASIBLE, DENY ACCESS TO AND USE OF THE
-              INTERFACE (INCLUDING BLOCKING CERTAIN ADDRESSES), TO ANY PERSON FOR ANY
-              REASON OR FOR NO REASON, INCLUDING WITHOUT LIMITATION FOR BREACH OF ANY
-              REPRESENTATION, WARRANTY, OR COVENANT CONTAINED IN THESE TERMS OF USE OR OF
-              ANY APPLICABLE LAW OR REGULATION. WE MAY TERMINATE YOUR USE OF THE
-              INTERFACE, WITHOUT WARNING, IN OUR SOLE DISCRETION.
-            </p>
-            <h3>8. Modifications and Interruptions</h3>
-            <p>
-              We reserve the right to change, modify, or remove the contents of the Site
-              at any time or for any reason at our sole discretion without notice.
-              However, we have no obligation to update any information on our Site.
-            </p>
-            <p>
-              We also reserve the right to modify or discontinue the Interface without
-              notice at any time. We will not be liable to you or any third party for any
-              modification, price change, suspension, or discontinuance of the Interface.
-            </p>
-            <p>
-              The Smart Contracts run autonomously on the blockchain and the operation of
-              the blockchain is not performed by us, but by the blockchain validator
-              nodes. We will not be liable to you or any third party for any problems in
-              the underlying blockchain, including without limitation: changes in
-              blockchain transaction fees, discontinuance of the blockchain, forks in the
-              blockchain, blockchain congestion, frontrunning and 51% attacks.
-            </p>
-            <p>
-              Djed depends on oracles that provide price data to the blockchain. We do not
-              operate the oracles. We will not be liable to you or any third party for any
-              problems in the oracles, including without limitation: oracle delays,
-              unavailability of the oracle, wrong data provided by the oracle.
-            </p>
-            <p>
-              We do not guarantee Djed will be available at all times. We may experience
-              hardware, software, or other problems or need to perform maintenance related
-              to Djed, resulting in interruptions, delays, or errors. We reserve the right
-              to change, revise, update, suspend, discontinue, or otherwise modify Djed at
-              any time or for any reason without notice to you. You agree that we have no
-              liability whatsoever for any loss, damage, or inconvenience caused by your
-              inability to access or use Djed during any downtime or discontinuance of the
-              Site. Nothing in these Terms of Use will be construed to obligate us to
-              maintain and support Djed or to supply any corrections, updates, or releases
-              in connection therewith.
-            </p>
-            <h3>9. Dispute Resolution</h3>
-            <p>
-              You hereby accept the non-territorial self-sovereign and autonomous nature
-              of Djed and waive any and all rights to, and hereby covenants not to, bring
-              any lawsuit, arbitration or other proceeding in any jurisdiction, judicial
-              body or forum arising under or relating to these Terms of Use or the use of
-              Djed.
-            </p>
-            <h3>10. Corrections</h3>
-            <p>
-              There may be information on the Site that contains typographical errors,
-              inaccuracies, or omissions, including descriptions, pricing, availability,
-              and various other information. We reserve the right to correct any errors,
-              inaccuracies, or omissions and to change or update the information on the
-              Site at any time, without prior notice.
-            </p>
-            <h3>11. Disclaimer</h3>
-            <p>
-              DJED IS PROVIDED ON AN AS-IS AND AS-AVAILABLE BASIS. YOU AGREE THAT YOUR USE
-              OF DJED WILL BE AT YOUR SOLE RISK. TO THE FULLEST EXTENT PERMITTED BY LAW,
-              WE DISCLAIM ALL WARRANTIES, EXPRESS OR IMPLIED, IN CONNECTION WITH DJED AND
-              YOUR USE THEREOF, INCLUDING, WITHOUT LIMITATION, THE IMPLIED WARRANTIES OF
-              MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, AND NON-INFRINGEMENT. WE
-              MAKE NO WARRANTIES OR REPRESENTATIONS ABOUT THE ACCURACY OR COMPLETENESS OF
-              THE SITE’S CONTENT OR THE CONTENT OF ANY WEBSITES LINKED TO THE SITE AND WE
-              WILL ASSUME NO LIABILITY OR RESPONSIBILITY FOR ANY (1) ERRORS, MISTAKES, OR
-              INACCURACIES OF CONTENT AND MATERIALS, (2) PERSONAL INJURY OR PROPERTY
-              DAMAGE, OF ANY NATURE WHATSOEVER, RESULTING FROM YOUR ACCESS TO AND USE OF
-              DJED, (3) ANY UNAUTHORIZED ACCESS TO OR USE OF OUR SECURE SERVERS AND/OR ANY
-              AND ALL PERSONAL INFORMATION AND/OR FINANCIAL INFORMATION STORED THEREIN,
-              (4) ANY INTERRUPTION OR CESSATION OF TRANSMISSION TO OR FROM THE SITE, (5)
-              ANY BUGS, VIRUSES, TROJAN HORSES, OR THE LIKE WHICH MAY BE TRANSMITTED TO OR
-              THROUGH DJED BY ANY THIRD PARTY, AND/OR (6) ANY ERRORS OR OMISSIONS IN ANY
-              CONTENT AND MATERIALS OR FOR ANY LOSS OR DAMAGE OF ANY KIND INCURRED AS A
-              RESULT OF THE USE OF ANY CONTENT POSTED, TRANSMITTED, OR OTHERWISE MADE
-              AVAILABLE VIA THE SITE. WE DO NOT WARRANT, ENDORSE, GUARANTEE, OR ASSUME
-              RESPONSIBILITY FOR ANY PRODUCT OR SERVICE ADVERTISED OR OFFERED BY A THIRD
-              PARTY THROUGH THE SITE, ANY HYPERLINKED WEBSITE, OR ANY WEBSITE OR MOBILE
-              APPLICATION FEATURED IN ANY BANNER OR OTHER ADVERTISING, AND WE WILL NOT BE
-              A PARTY TO OR IN ANY WAY BE RESPONSIBLE FOR MONITORING ANY TRANSACTION
-              BETWEEN YOU AND ANY THIRD-PARTY PROVIDERS OF PRODUCTS OR SERVICES. AS WITH
-              THE PURCHASE OF A PRODUCT OR SERVICE THROUGH ANY MEDIUM OR IN ANY
-              ENVIRONMENT, YOU SHOULD USE YOUR BEST JUDGMENT AND EXERCISE CAUTION WHERE
-              APPROPRIATE.
-            </p>
-            <h3>12. Limitations of Liability</h3>
-            <p>
-              IN NO EVENT WILL WE BE LIABLE TO YOU OR ANY THIRD PARTY FOR ANY DIRECT,
-              INDIRECT, CONSEQUENTIAL, EXEMPLARY, INCIDENTAL, SPECIAL, OR PUNITIVE
-              DAMAGES, INCLUDING LOST PROFIT, LOST REVENUE, LOSS OF DATA, OR OTHER DAMAGES
-              ARISING FROM YOUR USE OF DJED, EVEN IF WE HAVE BEEN ADVISED OF THE
-              POSSIBILITY OF SUCH DAMAGES. NOTWITHSTANDING ANYTHING TO THE CONTRARY
-              CONTAINED HEREIN, OUR LIABILITY TO YOU FOR ANY CAUSE WHATSOEVER AND
-              REGARDLESS OF THE FORM OF THE ACTION, WILL AT ALL TIMES BE LIMITED TO $0.00
-              USD.
-            </p>
-            <h3>13. Indemnification</h3>
-            <p>
-              You agree to defend, indemnify, and hold us harmless, including our
-              subsidiaries, affiliates, and all of our respective officers, agents,
-              partners, and employees, from and against any loss, damage, liability,
-              claim, or demand, including reasonable attorneys’ fees and expenses, made by
-              any third party due to or arising out of: (1) use of Djed; (2) breach of
-              these Terms of Use; (3) any breach of your representations and warranties
-              set forth in these Terms of Use; (4) your violation of the rights of a third
-              party, including but not limited to intellectual property rights; or (5) any
-              overt harmful act toward any other user of Djed with whom you connected via
-              your use of Djed. Notwithstanding the foregoing, we reserve the right, at
-              your expense, to assume the exclusive defense and control of any matter for
-              which you are required to indemnify us, and you agree to cooperate, at your
-              expense, with our defense of such claims. We will use reasonable efforts to
-              notify you of any such claim, action, or proceeding which is subject to this
-              indemnification upon becoming aware of it.
-            </p>
-            <h3>14. User Data</h3>
-            <p>
-              We will maintain no data about your use of Djed. Your use of Djed may cause
-              transactions to be registered on a blockchain. You agree that we shall have
-              no liability to you for any loss or corruption of any such data, and you
-              hereby waive any right of action against us arising from any such loss or
-              corruption of such data.
-            </p>
-            <h3>15. Electronic Communications, Transactions, and Signatures</h3>
-            <p>
-              Using Djed, sending us emails or messages through Discord, and completing
-              online forms constitute electronic communications. You consent to receive
-              electronic communications, and you agree that all agreements, notices,
-              disclosures, and other communications we provide to you electronically, via
-              email and on Djed, satisfy any legal requirement that such communication be
-              in writing. YOU HEREBY AGREE TO THE USE OF ELECTRONIC SIGNATURES, CONTRACTS,
-              ORDERS, AND OTHER RECORDS, AND TO ELECTRONIC DELIVERY OF NOTICES, POLICIES,
-              AND RECORDS OF TRANSACTIONS INITIATED OR COMPLETED BY US OR VIA THE USE OF
-              DJED. You hereby waive any rights or requirements under any statutes,
-              regulations, rules, ordinances, or other laws in any jurisdiction which
-              require an original signature or delivery or retention of non-electronic
-              records, or to payments or the granting of credits by any means other than
-              electronic means.
-            </p>
-            <h3>16. Miscellaneous</h3>
-            <p>
-              These Terms of Use and any policies or operating rules posted by us on the
-              Site or in respect to Djed constitute the entire agreement and understanding
-              between you and us. Our failure to exercise or enforce any right or
-              provision of these Terms of Use shall not operate as a waiver of such right
-              or provision. These Terms of Use operate to the fullest extent permissible
-              by law. We may assign any or all of our rights and obligations to others at
-              any time. We shall not be responsible or liable for any loss, damage, delay,
-              or failure to act caused by any cause beyond our reasonable control. If any
-              provision or part of a provision of these Terms of Use is determined to be
-              unlawful, void, or unenforceable, that provision or part of the provision is
-              deemed severable from these Terms of Use and does not affect the validity
-              and enforceability of any remaining provisions. There is no joint venture,
-              partnership, employment or agency relationship created between you and us as
-              a result of these Terms of Use or use of Djed. You agree that these Terms of
-              Use will not be construed against us by virtue of having drafted them. You
-              hereby waive any and all defenses you may have based on the electronic form
-              of these Terms of Use and the lack of signing by the parties hereto to
-              execute these Terms of Use.
-            </p>
-          </div>
-        </div>
-      </div>
-    </main>
-  );
-}
+const TermsOfUse = () => {
+  const [Terms, setTerms] = useState(null);
+
+  useEffect(() => {
+    const loadTerms = async () => {
+      try {
+        const termsPath = `./${process.env.REACT_APP_TERMS_OF_USE || "terms-of-use-default"}`;
+        const module = await import(`${termsPath}`);
+        setTerms(() => module.default);
+      } catch (error) {
+        console.error('Failed to load the terms:', error);
+      }
+    };
+    loadTerms();
+  }, []);
+
+  if (!Terms) {
+    return <div>Loading...</div>;
+  }
+
+  return <Terms />;
+};
+
+export default TermsOfUse;
